Bullet(총알) 최대 사거리 자료형 수정

+++ src/main/java/com/dace/dmgr/combat/interaction/ProjectileOption.java
@@ -18,7 +18,7 @@ public final class ProjectileOption {
     final int trailInterval = 14;
     /** 총알의 최대 사거리. (단위: 블록) */
     @Builder.Default
-    final int maxDistance = 70;
+    final double maxDistance = 70;
     /** 투사체가 유지되는 시간 (tick). {@code -1}로 설정 시 무한 지속 */
     @Builder.Default
     final long duration = -1;
